Add files via upload
Input
Each test contains multiple test cases. The first line contains the number of test cases 𝑇 (1≤𝑇≤10). Then the descriptions of the test cases follow.

The first line of each test case contains two integers 𝑛, 𝑚 (2≤𝑛≤1000, 1≤𝑚≤𝑛) — the number of people living in Hanh's apartment and the number of steel chains that the landlord requires, respectively.

The second line of each test case contains 𝑛 integers 𝑎1,𝑎2,…,𝑎𝑛 (0≤𝑎𝑖≤104) — weights of all fridges.

Output
For each test case:

If there is no solution, print a single integer −1.
Otherwise, print a single integer 𝑐 — the minimum total cost. The 𝑖-th of the next 𝑚 lines contains two integers 𝑢𝑖 and 𝑣𝑖 (1≤𝑢𝑖,𝑣𝑖≤𝑛, 𝑢𝑖≠𝑣𝑖), meaning that the 𝑖-th steel chain connects fridges 𝑢𝑖 and 𝑣𝑖. An arbitrary number of chains can be between a pair of fridges.
If there are multiple answers, print any.